:wrench: (chore): fix to sonarCloud, now i created a new controller to admin table
